% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_chr}
\alias{add_chr}
\title{Adding "chr" if it's not present}
\usage{
add_chr(chrs)
}
\arguments{
\item{chrs}{Chromosome or cluster names}
}
\value{
Data.frame with cluster ids and genes separated by commas
}
\description{
Adding "chr" if it's not present
}
