% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_splicing.R
\name{leaf_cutter_effect_sizes}
\alias{leaf_cutter_effect_sizes}
\title{Per intron effect sizes}
\usage{
leaf_cutter_effect_sizes(results)
}
\arguments{
\item{results}{From \code{\link{differential_splicing}}}
}
\value{
data.frame with a row for every tested intron and columns: intron, log effect size, baseline proportions, proportions in the second condition, and resulting deltaPSI
}
\description{
Per intron effect sizes
}
