% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leafcutter-package.R
\docType{package}
\name{leafcutter-package}
\alias{leafcutter-package}
\alias{leafcutter}
\title{The 'leafcutter' package.}
\description{
A DESCRIPTION OF THE PACKAGE
}
\references{
Stan Development Team (2018). RStan: the R interface to Stan. R package version 2.17.4. http://mc-stan.org
}
